microservices/admin/requirements.txt
asgiref==3.7.2
autopep8==1.6.0
Django==4.2.3
djangorestframework==3.14.0
pika==1.3.2
psycopg2==2.9.6
pycodestyle==2.8.0
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
